% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_base.R
\name{hello_base}
\alias{hello_base}
\title{Base function}
\usage{
hello_base(x = 2)
}
\arguments{
\item{x}{A parameter}
}
\value{
Some value
}
\description{
Base function
}
\details{
Some code with weird symbols: \code{pl$when(condition)} and \code{pl$then(output)}

Some equations: \eqn{\partial Y / \partial X = a + \varepsilon/2}
}
\examples{
hello_base()

mtcars$drat <- mtcars$drat + 1
head(mtcars[["drat"]], 2)
}
\references{
Ihaka R, Gentleman R (1996).
R: A Language for Data Analysis and Graphics.
\emph{Journal of Computational and Graphical Statistics}. \bold{5}(3), 299--314.
\doi{10.2307/139080}
}
\seealso{
\code{\link[base]{print}}, \code{\link{hello_r6}}
}
